Add driver simulator
This is a first version of the driver simulator.

Inside of the driver-simulator folder you will find:
- a REDME file to explain how to use the simulator;
- a simulate.js which implements the simulator;
- a vehicles.json used to simulate the vehicles

The simulator is capable of register a vehicle, update its location and
de-register the vehicle when it reaches its destination.

Improvements to this solution would include a docker configuration in
order to simplify the setup, and add more vehicles with new routes to
the JSON file.